+++ x/incentives/types/query.pb.gw.go
@@ -157,7 +157,7 @@ func request_Query_EpochIncentive_0(ctx context.Context, marshaler runtime.Marsh
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "epoch")
 	}
 
-	protoReq.Epoch, err = runtime.Int32(val)
+	protoReq.Epoch, err = runtime.Int64(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "epoch", err)
@@ -184,7 +184,7 @@ func local_request_Query_EpochIncentive_0(ctx context.Context, marshaler runtime
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "epoch")
 	}
 
-	protoReq.Epoch, err = runtime.Int32(val)
+	protoReq.Epoch, err = runtime.Int64(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "epoch", err)
